+++ src/main/java/com/ddiring/backend_market/investment/dto/response/InvestmentResponse.java
@@ -16,7 +16,7 @@ public class InvestmentResponse {
     private Integer investmentSeq;
     private String userSeq;
     private String projectId;
-    private Long investedPrice;
+    private Integer investedPrice;
     private Integer tokenQuantity;
     private String invStatus;
     private LocalDateTime investedAt;
